refactors Belt to break out UIKit functionality and properties into BeltBuilder.swift and simplifies belt construction logic there
